requirements.txt
aiohttp==3.7.4.post0
aiosignal==1.2.0
async-timeout==3.0.1
asyncpg==0.25.0
attrs==21.4.0
decouple==0.0.7
discord==1.7.3
discord-pretty-help==1.3.4
discord.py==1.7.3
disrank==0.0.2
frozenlist==1.3.0
Pillow==9.1.0
pynput==1.7.6
python-dotenv==0.20.0
requests==2.27.1
typing_extensions==4.2.0
urllib3==1.26.9
yarl==1.7.2